statement B/requirements.txt
absl-py
aioquic
altair
annotated-types
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
astunparse
async-lru
attrs
Babel
beautifulsoup4
bleach
blinker
Brotli
browsermob-proxy
cachetools
certifi
cffi
charset-normalizer
click
comm
contextlib2
contourpy
cryptography
cycler
Cython
debugpy
decorator
defusedxml
distro
executing
faiss-cpu
fastjsonschema
filelock
Flask
Flask-Cors
flatbuffers
fonttools
fqdn
fsspec
future
gast
gensim
gitdb
GitPython
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
gpg
grpcio
h11
h2
h5py
hpack
httpcore
httpx
huggingface-hub
hyperframe
idna
iniconfig
innvestigate
ipykernel
ipython
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
jiter
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kaitaistruct
keras
Keras-Applications
Keras-Preprocessing
keras-tuner
kiwisolver
kt-legacy
ldap3
libclang
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
mitmproxy
mitmproxy-macos
mitmproxy_rs
ml-dtypes
mpmath
msgpack
namex
narwhals
nbclient
nbconvert
nbformat
neo4j
nest-asyncio
networkx
nltk
notebook
notebook_shim
numpy
openai
opencv-python
opt-einsum
optree
outcome
overrides
packaging
pandas
pandocfilters
parso
passlib
pdf2image
pexpect
pillow
platformdirs
pluggy
POT
prometheus_client
prompt_toolkit
proto-plus
protobuf
psutil
ptyprocess
publicsuffix2
pure-eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydeck
Pygments
pylsqpack
PyMuPDF
PyMuPDFb
pyOpenSSL
pyparsing
pyperclip
PySocks
pytesseract
pytest
python-dateutil
python-docx
python-dotenv
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
randomsearch
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rich
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
safetensors
scikit-learn
scipy
selenium
Send2Trash
sentence-transformers
service-identity
setuptools
six
smart-open
smmap
sniffio
sortedcontainers
soupsieve
stack-data
streamlit
stripe
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
termcolor
terminado
tf-slim
tf_keras
threadpoolctl
tinycss2
tokenizers
toml
torch
tornado
tqdm
traitlets
transformers
trio
trio-websocket
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
urwid
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wheel @ file:///opt/homebrew/Cellar/python%403.12/3.12.5/libexec/wheel-0.44.0-py3-none-any.whl#sha256=8d8a590e50bda8422000110e90bc730ebfcd8363d2de80793ac819b7f252e97d
widgetsnbextension
wrapt
wsproto
zstandard
